18/5/7
服务器部署完成 开始线上测试

+++ app/Issue.php
@@ -26,7 +26,7 @@
 
     function issue(){
         
-        if(!isset($_POST['qd'],$_POST['zd'],$_POST['qdName'],$_POST['zdName'],$_POST['type'],$_POST['cf'])){
+        if(!isset($_POST['qd'],$_POST['zd'],$_POST['qdName'],$_POST['zdName'],$_POST['type'],$_POST['cf'],$_POST['wxformId'])){
             $info = new IssueErrorInfo2("参数非法");
             Tools::infoBack($info);
             return;
@@ -38,8 +38,8 @@
         $bz = $_POST['bz'];
         $type = $_POST['type'];
         $cf = $_POST['cf'];
-        
-        
+        $wxformId = $_POST['wxformId'];
+
         // $qd = array('lat'=>39.071510,'lng'=>117.190091);        
         // $zd = array('lat'=>39.071510,'lng'=>117.190010);
         // $bz = "test";
@@ -49,9 +49,11 @@
 
         $databaseTools = new databaseTools();
         $database = $databaseTools->databaseInit();
-        if($database->select('user_basic','contribution',array(
+        $gxz = $database->select('user_basic','contribution',array(
             'UID'=>$masterUID
-        )) < 0){
+        ));
+
+        if(count($gxz) == 0 || $gxz[0] < 0){
             $info = new IssueErrorInfo2("贡献值不足");
             Tools::infoBack($info);
             return;
@@ -60,8 +62,14 @@
         $qd = json_decode($qd,true);
         $zd = json_decode($zd,true);
 
-        
         $fq = time();
+        if($fq >= $cf)
+        {
+            $info = new IssueErrorInfo2("出发时间不能早于当前");
+            Tools::infoBack($info);
+            return;
+        }
+
         $key = "".$masterUID.$fq;
         $formID = Particle::timeFromParticle(Particle::generateParticle($key));
         $flag = 0;
@@ -75,7 +83,7 @@
             return;
         }
         //待查数据格式
-        $distance = $result['result']['elements'][0]['distance'];
+        $distance = $result['result']['elements'][0]['distance'] / 1000;
 
         // $distance = 1000.0;
         $datas = array(
@@ -90,17 +98,19 @@
             'cf'=>$cf,
             'distance'=>$distance,
             'flag'=>$flag,
-            'type'=>$type
+            'type'=>$type,
+            'wxformId'=>$wxformId
         );
         // --------------------------------------------------
         //基本表插入
 
         $result = $database->insert('form_basic',$datas);
-        if($result == false){
-            $info = new IssueErrorInfo2("服务器错误 请重新发布订单");
-            Tools::infoBack($info);
-            return; 
-        }
+
+//        if($result == false){
+//            $info = new IssueErrorInfo2("服务器错误 请重新发布订单");
+//            Tools::infoBack($info);
+//            return;
+//        }
 
         //插入待服务列表
         $geohash = new Geohash();
@@ -113,13 +123,11 @@
             'start_geohash'=>$start_geohash,
             'end_geohash'=>$end_geohas
         );
+
         $result = $database->insert('waiting_server',$datas);
 
         $info = new IssueErrorInfo0($formID);
-        Tools::infoBack($info); 
-        
-        
-    
+        Tools::infoBack($info);
     }
 
     function quitIssue(){
@@ -132,13 +140,24 @@
         $databaseTools = new databaseTools();
         $database = $databaseTools->databaseInit();
 
-        $result = $database->select('waiting_server','form_id',array('form_id'=>$formID));
+        $result = $database->select('form_basic',array('formID', 'flag', 'masterUID', 'serverUID'),array('formID'=>$formID));
         if(empty($result)){
             $info = new IssueErrorInfo2("该订单不存在或已经结束");
             Tools::infoBack($info); 
         }else{
-            //从等待列表中删除
-            $result = $database->delete('waiting_server',array('form_id'=>$formID));
+			$result = $result[0];
+            if($result['masterUID']!=$_SESSION['UID']&&$_SESSION['UID']!=$result['serverUID']){
+                $info = new errorInfo("该用户没有权限");
+                Tools::infoBack($info);
+                return;
+            }
+            if ($result['flag']==0)
+            {
+                //从等待列表中删除
+                $result = $database->delete('waiting_server',array('form_id'=>$formID));
+            }else{
+                $result = true;
+            }
             if($result){
                 //更新基础表中的flag
                 $database->update('form_basic',array('flag'=>5,'bz[+]'=>$bz),array('formID'=>$formID));
